requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
python-dotenv==1.0.1
python-telegram-bot==21.0.1
gunicorn==22.0.0
Werkzeug==3.0.1